pyproject.toml
[project]
name = 'radpy'
version = "0.1.0"
description = "Robust Angular Diameters in Python: an angular diameter fitting routine for multi-wavelength interferometric data"
authors = [
    {name = "Ashley Elliott", email = "elliottashleya99@gmail.com"}
]
license = {text = "GPL-3.0"}
readme = "README.md"
requires-python = ">=3.10,<3.12"
keywords = ["interferometry", "angular diameters", "CHARA"]
dependencies = [
    "astropy~=6.1.6",
    "corner~=2.2.2",
    "ipympl~=0.9.4",
    "lmfit~=1.2.2",
    "matplotlib~=3.9.2",
    "numpy~=1.25.2",
    "pandas~=2.2.2",
    "scipy~=1.10.0",
    "gaiadr3-zeropoint (>=0.0.4,<0.1.0)",
    "pytest~=7.1.2"
]
[project.urls]
repository = "https://github.com/spaceashley/radpy.git"

[tool.poetry.dependencies]
astropy = "6.1.6"
corner = "2.2.2"
ipympl = "0.9.4"
lmfit = "1.2.2"
matplotlib = "3.9.2"
numpy = "1.25.2"
pandas = "^2.2.2"
scipy = "1.10.0"
#gaiadr3-zeropoint = "0.1.0"
pytest = "7.1.2"


#[tool.poetry.build]
#include = ["src/radpy/data/ldc_I_table/csv",
#           "src/radpy/data/ldc_k_table.csv",
#           "src/radpy/data/ldctable.csv"
#]

[build-system]
requires = [
    "poetry-core"
]
build-backend = "poetry.core.masontry.api"